Moved chat from gameboard to ChatFragment
The keyboard closes and text view clears when you press the send button.
