+++ app/src/main/java/com/example/ryanblaser/tickettoride/GUI/Activities/WaitingActivity.java
@@ -44,17 +44,17 @@ public class WaitingActivity extends AppCompatActivity {
         button_start_game.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                if (isAtLeastTwoPlayers()) {
+//                if (isAtLeastTwoPlayers()) {
                     int gameId = ClientFacade.SINGLETON.getClientModel().getInt_curr_gameId();
                     int playerSize = ClientFacade.SINGLETON.getClientModel().getGameId_to_usernames().get(gameId).size();
                     List<String> usernamesInGame = ClientFacade.SINGLETON.getClientModel().getGameId_to_usernames().get(gameId);
 
                     ClientFacade.SINGLETON.startGame(gameId, usernamesInGame);
                     Toast.makeText(getBaseContext(), "Starting Game with " + playerSize + " players!", Toast.LENGTH_SHORT).show();
-                }
-                else {
-                    Toast.makeText(getBaseContext(), "Need 2-5 players to start the game", Toast.LENGTH_SHORT).show();
-                }
+//                }
+//                else {
+//                    Toast.makeText(getBaseContext(), "Need 2-5 players to start the game", Toast.LENGTH_SHORT).show();
+//                } //TODO: Uncomment to actually play game
 
             }
         });
