+++ app/src/main/res/layout/fragment_chat.xml
@@ -12,24 +12,26 @@
         android:layout_alignParentTop="true"
         android:layout_above="@+id/send" />
 
-    <!--<EditText-->
-        <!--android:id="@+id/text_message"-->
-        <!--android:layout_width="500dp"-->
-        <!--android:layout_height="35dp"-->
-        <!--android:layout_alignParentBottom="true"-->
-        <!--android:layout_alignParentStart="true"-->
-        <!--android:background="@color/black_overlay"-->
-        <!--android:textColor="@color/colorAccent" />-->
+    <EditText
+        android:id="@+id/text_message"
+        android:layout_width="500dp"
+        android:layout_height="35dp"
+        android:layout_alignParentBottom="true"
+        android:layout_alignParentStart="true"
+        android:background="@color/black_overlay"
+        android:hint="Tap to send a message"
+        android:textColor="@color/colorAccent"
+        android:layout_alignParentEnd="true" />
 
-    <!--<Button-->
-        <!--android:id="@+id/send"-->
-        <!--style="@style/Widget.AppCompat.Button.Colored"-->
-        <!--android:layout_width="wrap_content"-->
-        <!--android:layout_height="wrap_content"-->
-        <!--android:elevation="0dp"-->
-        <!--android:text="Send"-->
-        <!--android:textSize="12sp"-->
-        <!--android:layout_alignTop="@+id/text_message"-->
-        <!--android:layout_alignParentEnd="true" />-->
+    <Button
+        android:id="@+id/send"
+        style="@style/Widget.AppCompat.Button.Colored"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:elevation="0dp"
+        android:text="Send"
+        android:textSize="12sp"
+        android:layout_alignTop="@+id/text_message"
+        android:layout_alignParentEnd="true" />
 
 </RelativeLayout>
